experiments/2_e_conversational_analytics/requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# AI and ML
crewai>=0.1.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5

# Speech recognition
SpeechRecognition>=3.8.1
# PyAudio>=0.2.11  # Requires system dependencies: sudo apt install portaudio19-dev python3-dev

# Data processing
openpyxl>=3.1.0
xlrd>=2.0.1

# Utilities
python-dotenv>=1.0.0
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

